requirements.txt
argon2-cffi>=23.3.0
cryptography>=41.0.0
pyperclip>=1.8.2
python-dotenv>=1.0.0

pip install -r requirements.txt.